pyproject.toml
[project]
name = "website"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask-sqlalchemy>=3.1.1",
    "flask>=3.1.2",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "alembic>=1.16.5",
]
